requirements.txt
streamlit
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
transformers
timm
opencv-python
matplotlib
pytorch-grad-cam